backend/requirements.txt
fastapi>=0.109.0,<0.110.0
uvicorn>=0.27.0,<0.28.0
sqlalchemy>=1.4.0,<1.5.0
psycopg2-binary>=2.9.1,<3.0.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.1.0
email-validator>=1.1.3,<1.2.0
redis>=4.0.0,<4.1.0
pydantic>=2.6.0,<3.0.0
python-dotenv>=0.19.0,<0.20.0
requests>=2.26.0,<2.27.0
PyJWT>=2.1.0,<2.2.0